Relocate createMetricSpec and related test cases
