+++ pkg/scaling/utils/metrics.go
@@ -1,6 +1,9 @@
 package utils
 
-import v2 "k8s.io/api/autoscaling/v2"
+import (
+	v2 "k8s.io/api/autoscaling/v2"
+	"k8s.io/apimachinery/pkg/api/resource"
+)
 
 func GetTargetAverageValue(metricSpecs []v2.MetricSpec) float64 {
 	var totalAverageValue float64
@@ -19,4 +22,18 @@ func GetTargetAverageValue(metricSpecs []v2.MetricSpec) float64 {
 		return totalAverageValue / count
 	}
 	return 0
+}
+
+func CreateMetricSpec(averageValue int64, metricName string) v2.MetricSpec {
+	qty := resource.NewQuantity(averageValue, resource.DecimalSI)
+	return v2.MetricSpec{
+		External: &v2.ExternalMetricSource{
+			Target: v2.MetricTarget{
+				AverageValue: qty,
+			},
+			Metric: v2.MetricIdentifier{
+				Name: metricName,
+			},
+		},
+	}
 }
